forth merge

+++ src/api/routes.py
@@ -82,6 +82,9 @@ def get_user_information():
 @api.route('/users/<int:user_id>', methods=['DELETE'])
 @jwt_required()
 def delete_user(user_id):
+    token_user_id = get_jwt_identity()
+    if token_user_id != user_id:
+        return jsonify({"Not allowed to delete"}), 403
     user = User.query.get(user_id)
     if user:
         db.session.delete(user)
@@ -89,6 +92,7 @@ def delete_user(user_id):
         return jsonify({"User deleted successfully"}), 200
     else:
         return jsonify({"User not found"}), 404
+    
 # GET all books
 @api.route('/books', methods=['GET'])
 @jwt_required()
